refactor(header): rename file header to NodeHeader
